--- MoltenVKShaderConverter/MoltenVKGLSLToSPIRVConverter.xcframework
@@ -1,1 +1,0 @@
-../Package/Latest/MoltenVKShaderConverter/MoltenVKGLSLToSPIRVConverter.xcframework
