--- MoltenVKShaderConverter/MoltenVKSPIRVToMSLConverter.xcframework
@@ -1,1 +1,0 @@
-../Package/Latest/MoltenVKShaderConverter/MoltenVKSPIRVToMSLConverter.xcframework
