--- MoltenVKShaderConverter/include/MoltenVKGLSLToSPIRVConverter/GLSLConversion.h
@@ -1,1 +1,0 @@
-../../MoltenVKGLSLToSPIRVConverter/GLSLConversion.h
