--- MoltenVKShaderConverter/include/MoltenVKGLSLToSPIRVConverter/GLSLToSPIRVConverter.h
@@ -1,1 +1,0 @@
-../../MoltenVKGLSLToSPIRVConverter/GLSLToSPIRVConverter.h
