--- MoltenVKShaderConverter/include/MoltenVKSPIRVToMSLConverter/SPIRVToMSLConverter.h
@@ -1,1 +1,0 @@
-../../MoltenVKSPIRVToMSLConverter/SPIRVToMSLConverter.h
