+++ Scripts/package_shader_converter_xcframework.sh
@@ -11,15 +11,12 @@ export MVK_XCFWK_DEST_DIR="${PROJECT_DIR}/Package/${CONFIGURATION}/${MVK_PROD_BA
 # Assemble the headers for the shader frameworks
 hdr_dir="${MVK_XCFWK_STAGING_DIR}/Headers"
 mkdir -p "${hdr_dir}"
-rm -rf "${hdr_dir}/MoltenVKSPIRVToMSLConverter"
-cp -pRL "${PROJECT_DIR}/${MVK_PROD_BASE_NAME}/include/MoltenVKSPIRVToMSLConverter" "${hdr_dir}"
-rm -rf "${hdr_dir}/MoltenVKGLSLToSPIRVConverter"
-cp -pRL "${PROJECT_DIR}/${MVK_PROD_BASE_NAME}/include/MoltenVKGLSLToSPIRVConverter" "${hdr_dir}"
+rm -rf "${hdr_dir}/MoltenVKShaderConverter"
+cp -pRL "${PROJECT_DIR}/${MVK_PROD_BASE_NAME}/include/MoltenVKShaderConverter" "${hdr_dir}"
 
 # Also copy headers to an include directory in the package.
 # This will not be needed once the XCFramework can be created with a Headers directory.
 mkdir -p "${MVK_XCFWK_DEST_DIR}"
 cp -pRL "${PROJECT_DIR}/${MVK_PROD_BASE_NAME}/include/" "${MVK_XCFWK_DEST_DIR}/include"
 
-create_xcframework "MoltenVKSPIRVToMSLConverter"
-create_xcframework "MoltenVKGLSLToSPIRVConverter"
+create_xcframework "MoltenVKShaderConverter"
